+++ CHANGELOG.md
@@ -3,6 +3,7 @@ FEATURES:
 * **New Resource:** `yandex_datatransfer_endpoint`
 * **New Resource:** `yandex_datatransfer_transfer`
 * managed-kubernetes: add `container_runtime` attribute to `yandex_kubernetes_node_group` resource and data source
+* managed-elasticsearch: add `maintenance_window` section in `yandex_mdb_elasticsearch_cluster` resource and data source
 
 ## 0.69.0 (December 27, 2021)
 FEATURES: